pg_reader/requirements.txt
quixstreams==2.4.1
python-dotenv
psycopg2-binary